agents/agent_crawler/requirements.txt
crawl4ai
playwright
pydantic>=2.0
pydantic-ai
google-adk
uvicorn
fastapi